reverted DeltaTime back to float to provide backwards compat, but added DeltaTimeDouble for those who want the newly provided (by godot4) double precision :)

+++ demo/IconAnimation.cs
@@ -47,7 +47,7 @@ public partial class IconAnimation : Node2D
 
         while (angle < 2 * Mathf.Pi)
         {
-            angle += speed * (float)Co.DeltaTime;
+            angle += speed * Co.DeltaTime;
             Rotation = angle;
             yield return null;
         }
